fix timestamp by patching ElvUI instead
In HyperLinkedCPL colored links are replaced, so that | will be \10.
The problem: \10 is the ASCII control code for a line feed and
E:StripString() trims it away (^%s+), so my colored, linked timestamp
at the beginning of the message breaks the message.

The provided ElvUI diff patch changes the HyperLinkedCPL function, so
that it uses the reserved ASCII control codes "Device Control 1" (\17)
and "Device Control 2" (\18) instead, so they don't get trimmed.

The patch also includes a fix for mawpower links in chat not having
tooltips and makes |n available in unit/raid frame texts.

Move .patch file to your ElvUI repo and apply with
git apply ElvUI-12.79.patch